requirements.txt
pandas
streamlit
numpy
matplotlib
scipy
Pillow